working on an alternative way for cam positioning

+++ nova2d/nova-zero/src/core/Positional.cpp
@@ -92,8 +92,8 @@ namespace novazero
 			float x = (xR * Game::s_Width) - (m_Drawable->GetWidth() / 2);
 			float y = (yR * Game::s_Height) - (m_Drawable->GetHeight() / 2);
 						
-			SetOffsetX((int)x);
-			SetOffsetY((int)y);
+			SetOffsetX(x);
+			SetOffsetY(y);
 		}
 	}
 }
